FEAT: tokenizer
- modify behaviour of fn get_next_token:

BEFORE:	every quote-enclosed section is considered a token,
regardless if it is, or not, separated with space(s)
from the rest of the command

NOW: every quote-enclosed or non-enclosed section is considered a token,
only if it's separated by space(s) from the rest of the command,
regardless if a subsection is quote-enclosed.
Does not apply to operators!

- improve algorithm of fn tokenizer
- remove fn: check_quote, check_operator, extract_str
- update header
- create separate tester function for tokenizer

TLDR: argument distinction is now based only on space separation