UPDATE: Add to t_lst_str struct, complete export fn, change 'env_bak' to 'variables' and related changes for storage and extraction convinience
The struct t_lst_str gets two new fields key and val. This is for variables list (previously env_bak) to store the env variables as a key-value pair for ease of extraction later.
Also, update function copy_env to store env variables as key-val pairs.
Add function to store extra variables in the minishell memory created by a user, but not to put it in the shl->env
Add the export function that writes any new variables created using export