FIX/UPDATE: Updated functions ft_remove_node and ft_find_node
ft_remove_node function was added to remove a node from anywhere in a
list.
ft_find_node function has been modified to facilitate finding the node
even when the search string is a substring of the field being searched.
This is done for the fn to work with both shl->env and shl->variables.